Refactoring of services to prevent merge (big) conflicts
